fix: move track exception to new file
