+++ packages/hms-video-store/src/error/HMSException.ts
@@ -1,6 +1,5 @@
 import { HMSAction } from './HMSAction';
 import { IAnalyticsPropertiesProvider } from '../analytics/IAnalyticsPropertiesProvider';
-import { HMSTrackExceptionTrackType } from '../media/tracks/HMSTrackExceptionTrackType';
 import { HMSSignalMethod } from '../signal/jsonrpc/models';
 
 export class HMSException extends Error implements IAnalyticsPropertiesProvider {
@@ -46,39 +45,6 @@ export class HMSException extends Error implements IAnalyticsPropertiesProvider 
         description: ${this.description};
         isTerminal: ${this.isTerminal};
         nativeError: ${this.nativeError?.message};
-      }`;
-  }
-}
-
-export class HMSTrackException extends HMSException {
-  constructor(
-    code: number,
-    name: string,
-    action: HMSAction | HMSSignalMethod,
-    message: string,
-    description: string,
-    public trackType: HMSTrackExceptionTrackType,
-  ) {
-    super(code, name, action, message, description, false);
-  }
-
-  toAnalyticsProperties() {
-    return {
-      ...super.toAnalyticsProperties(),
-      track_type: this.trackType,
-    };
-  }
-
-  toString() {
-    return `{
-        code: ${this.code};
-        name: ${this.name};
-        action: ${this.action};
-        message: ${this.message};
-        description: ${this.description};
-        isTerminal: ${this.isTerminal};
-        nativeError: ${this.nativeError?.message};
-        trackType: ${this.trackType};
       }`;
   }
 }
